Move the MockGraphite class to its own file.
